Build Whonix on 02_install_groups_pre before any other packages installed to prevent package conflicts

+++ whonix-gateway/02_install_groups_pre.sh
@@ -1,6 +1,34 @@
 #!/bin/bash -e
 # vim: set ts=4 sw=4 sts=4 et :
 
+# ==============================================================================
+#                           WHONIX 10 - WIP NOTES
+# ==============================================================================
+# 
+# TODO - TEMP TEST:
+# ------------------------------------------------------------------------------
+# -
+#
+# TODO - EXPERIMENT:
+# ------------------------------------------------------------------------------
+# -
+#
+# TODO - FIX:
+# ------------------------------------------------------------------------------
+# - dialog boxes partial display as semi-transparent (wheezy + jessie)
+#   - test to see if that is still case with gnome enable workstation
+#   - possible QT or TrollTech.conf issue?
+#
+# WHONIX RELATED BUGS:
+# ------------------------------------------------------------------------------
+# -
+#
+# REPAIR AFTER SUCCESSFUL BUILDS:
+# ------------------------------------------------------------------------------
+# -
+# 
+# ==============================================================================
+
 source "${SCRIPTSDIR}/vars.sh"
 source "${SCRIPTSDIR}/distribution.sh"
 
@@ -20,15 +48,30 @@ if ! [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared_groups" ]; then
     info ' Installing extra packages in packages_whonix.list file'
     #### '----------------------------------------------------------------------
     installPackages packages_whonix.list
+
+    #### '----------------------------------------------------------------------
+    info ' Installing extra packages from Whonix 30_dependencies'
+    #### '----------------------------------------------------------------------
+    source "${WHONIX_DIR}/buildconfig.d/30_dependencies"
+    aptInstall ${whonix_build_script_build_dependency}
+
     touch "${INSTALLDIR}/${TMPDIR}/.whonix_prepared_groups"
 fi
 
 
-# ------------------------------------------------------------------------------
+# ==============================================================================
 # chroot Whonix build script
-# ------------------------------------------------------------------------------
+# ==============================================================================
 read -r -d '' WHONIX_BUILD_SCRIPT <<'EOF' || true
 ################################################################################
+# This script is executed from chroot most likely as the user `user`
+# 
+# - The purpose is to do a few pre-fixups that are directly related to whonix
+#   build process
+# - Then, finally, call `whonix_build_post` as sudo with a clean (no) ENV
+#
+################################################################################
+
 # Pre Fixups
 sudo mkdir -p /boot/grub2
 sudo touch /boot/grub2/grub.cfg
@@ -43,33 +86,55 @@ sudo apt-mark hold sysvinit
 sudo apt-mark hold grub-pc grub-pc-bin grub-common grub2-common
 
 # Whonix expects haveged to be started
+# ------------------------------------------------------------------------------
 sudo /etc/init.d/haveged start
-################################################################################
-# Whonix installation
-export WHONIX_BUILD_UNATTENDED_PKG_INSTALL="1"
 
-pushd ~/Whonix
-sudo ~/Whonix/whonix_build \
+# Use sudo with clean ENV to build Whonix; any ENV options will be set there
+# ------------------------------------------------------------------------------
+sudo ~/whonix_build_post $@
+EOF
+
+# ==============================================================================
+# chroot Whonix post build script
+# ==============================================================================
+read -r -d '' WHONIX_BUILD_SCRIPT_POST <<'EOF' || true
+#!/bin/bash -e
+# vim: set ts=4 sw=4 sts=4 et :
+
+# Prevents Whonix makefile use of shared memory 'sem_open: Permission denied'
+# ------------------------------------------------------------------------------
+echo tmpfs /dev/shm tmpfs defaults 0 0 >> /etc/fstab
+mount /dev/shm
+
+# =============================================================================
+# WHONIX BUILD COMMAND
+# =============================================================================
+pushd /home/user/Whonix
+/home/user/Whonix/whonix_build \
     --flavor $1 \
     -- \
     --build \
     --arch amd64 \
     --kernel linux-image-amd64 \
     --headers linux-headers-amd64 \
-    --target root \
     --freshness current \
+    --target root \
     --report minimal \
     --verifiable minimal \
+    --allow-uncommitted true \
+    --allow-untagged true \
     --sanity-tests false || { exit 1; }
 popd
 EOF
 
+
 # Some Additional Whonix build options
 # ====================================
-# --tb close  # Install tor-browser
-# --allow-uncommitted true
-# --allow-untagged true
-# --testing-frozen-sources  # Jessie; no current sources
+#    --tb close  # Install tor-browser \
+#    --allow-uncommitted true \
+#    --allow-untagged true \
+#    --testing-frozen-sources  # Jessie; no current sources \
+
 
 ##### '-------------------------------------------------------------------------
 debug ' Preparing Whonix for installation'
@@ -111,14 +176,27 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared_groups" ] && ! [ -f "${INSTALL
     #### '----------------------------------------------------------------------
     info ' Installing Whonix build scripts'
     #### '----------------------------------------------------------------------
-    echo "${WHONIX_BUILD_SCRIPT}" > "${INSTALLDIR}/home/user/whonix_build.sh"
-    chmod 0755 "${INSTALLDIR}/home/user/whonix_build.sh"
+    echo "${WHONIX_BUILD_SCRIPT_POST}" > "${INSTALLDIR}/home/user/whonix_build_post"
+    chmod 0755 "${INSTALLDIR}/home/user/whonix_build_post"
+
+    echo "${WHONIX_BUILD_SCRIPT}" > "${INSTALLDIR}/home/user/whonix_build"
+    chmod 0755 "${INSTALLDIR}/home/user/whonix_build"
 
     #### '----------------------------------------------------------------------
     info ' Removing apt-listchanges if it exists,so no prompts appear'
     #### '----------------------------------------------------------------------
     #      Whonix does not handle this properly, but aptInstall packages will
     aptRemove apt-listchanges || true
+
+    #### '----------------------------------------------------------------------
+    debug 'XXX: Whonix10/11 HACK'
+    #### '----------------------------------------------------------------------
+    rm -f "${INSTALLDIR}/etc/network/interfaces"
+    cat > "${INSTALLDIR}/etc/network/interfaces" <<'EOF'
+# interfaces(5) file used by ifup(8) and ifdown(8)
+# Include files from /etc/network/interfaces.d:
+source-directory /etc/network/interfaces.d
+EOF
 
     #### '----------------------------------------------------------------------
     info ' Copying additional files required for build'
@@ -137,6 +215,7 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared" ] && ! [ -f "${INSTALLDIR}/${
         chroot su user -c 'mkdir /home/user/Whonix'
     fi
 
+    # XXX: TODO: Need to get from ENV $WHONIX_DIR as we can't always be sure where dir is
     mount --bind "../Whonix" "${INSTALLDIR}/home/user/Whonix"
 
     if [ "${TEMPLATE_FLAVOR}" == "whonix-gateway" ]; then
@@ -154,7 +233,17 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared" ] && ! [ -f "${INSTALLDIR}/${
     mount --bind /dev "${INSTALLDIR}/dev"
     mount --bind /dev/pts "${INSTALLDIR}/dev/pts"
 
-    chroot su user -c "cd ~; ./whonix_build.sh ${BUILD_TYPE} ${DIST}" || { exit 1; }
+    # Enable for logging...
+    # XXX: Remove static reference and store in template-whonix $DIR
+    #BUILD_LOG=/home/user/qubes/qubes-src/template-whonix/whonix.log
+
+    info 'Executing whonix_build script now...'
+    if [ "x${BUILD_LOG}" != "x" ]; then
+        chroot su user -c "cd ~; ./whonix_build ${BUILD_TYPE} ${DIST}" 3>&2 2>&1 | tee -a ${BUILD_LOG} || { exit 1; }
+    else
+        chroot su user -c "cd ~; ./whonix_build ${BUILD_TYPE} ${DIST}" || { exit 1; }
+    fi
+
     touch "${INSTALLDIR}/${TMPDIR}/.whonix_installed"
 fi
 
@@ -169,19 +258,23 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_installed" ] && ! [ -f "${INSTALLDIR}/$
     #### '----------------------------------------------------------------------
     pushd "${INSTALLDIR}/etc/network"
     {
-        rm -f interfaces;
-        ln -s interfaces.backup interfaces;
+        if [ -e 'interfaces.backup' ]; then
+            rm -f interfaces;
+            ln -s interfaces.backup interfaces;
+        fi
     }
     popd
 
     #### '----------------------------------------------------------------------
-    info ' Temporarily retore original resolv.conf for remainder of install process'
-    info ' (Will be restored back in wheezy+whonix/04_qubes_install_post.sh)'
+    info ' Temporarily restore original resolv.conf for remainder of install process'
+    info ' (Will be restored back in jessie+whonix/04_qubes_install_post.sh)'
     #### '----------------------------------------------------------------------
     pushd "${INSTALLDIR}/etc"
     {
-        rm -f resolv.conf;
-        cp -p resolv.conf.backup resolv.conf;
+        if [ -e 'resolv.conf.backup' ]; then
+            rm -f resolv.conf;
+            cp -p resolv.conf.backup resolv.conf;
+        fi
     }
     popd
 
@@ -191,8 +284,10 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_installed" ] && ! [ -f "${INSTALLDIR}/$
     #### '----------------------------------------------------------------------
     pushd "${INSTALLDIR}/etc"
     {
-        rm -f hosts;
-        cp -p hosts.anondist-orig hosts;
+        if [ -e 'hosts.anondist-orig' ]; then
+            rm -f hosts;
+            cp -p hosts.anondist-orig hosts;
+        fi
     }
     popd
